+++ nn_dataflow/core/option.py
@@ -28,6 +28,7 @@ OPTION_LIST = ['sw_gbuf_bypass',
                'partition_dimensional',
                'partition_batch',
                'partition_ifmaps',
+               'opt_goal',
                'ntops',
                'nprocesses',
                'verbose',
@@ -64,6 +65,7 @@ class Option(namedtuple('Option', OPTION_LIST)):
         kwdict.setdefault('partition_dimensional', False)
         kwdict.setdefault('partition_batch', False)
         kwdict.setdefault('partition_ifmaps', False)
+        kwdict.setdefault('opt_goal', 'e')
         kwdict.setdefault('ntops', 1)
         kwdict.setdefault('nprocesses', 1)
         kwdict.setdefault('verbose', False)
@@ -81,6 +83,10 @@ class Option(namedtuple('Option', OPTION_LIST)):
         if ntp.partition_ifmaps and not ntp.partition_hybrid:
             raise ValueError('Option: partition_ifmaps requires '
                              'partition_hybrid to be set.')
+
+        if ntp.opt_goal not in ['e', 'd', 'ed']:
+            raise ValueError('Option: opt_goal is invalid, must be one of '
+                             '\'e\', \'d\', and \'ed\'.')
 
         return ntp